+++ package.json
@@ -50,6 +50,7 @@
     "react-datepicker": "^4.8.0",
     "react-dom": "18.2.0",
     "react-icons": "^4.4.0",
+    "react-leaflet": "^4.1.0",
     "react-player": "^2.10.1",
     "react-query": "^3.39.2",
     "react-redux": "^8.0.2",
@@ -75,10 +76,10 @@
     "@types/react-datepicker": "^4.4.2",
     "@types/react-dom": "18.0.6",
     "@types/react-slick": "^0.23.10",
-    "@types/sort-array": "^4.1.0",
     "@types/react-star-ratings": "^2.3.0",
     "@types/react-stars": "^2.2.1",
     "@types/react-vertical-timeline-component": "^3.3.3",
+    "@types/sort-array": "^4.1.0",
     "eslint": "8.23.1",
     "eslint-config-next": "12.3.0",
     "eslint-config-prettier": "^8.5.0",
