+++ src/components/SendWeb3Transaction.tsx
@@ -1,46 +1,25 @@
-import {
-  Box,
-  Button,
-  Container,
-  Flex,
-  FormControl,
-  Input,
-  Link,
-  Select,
-  Text,
-} from "@chakra-ui/react";
+import { Box, Button, Flex, FormControl, Input, Text } from "@chakra-ui/react";
 import { useAddRecentTransaction } from "@rainbow-me/rainbowkit";
 import { parseEther } from "ethers/lib/utils";
-import { useEffect, useState } from "react";
-import { useDebounce } from "use-debounce";
+import { useState } from "react";
 import {
   erc20ABI,
-  useContract,
   useContractWrite,
   usePrepareContractWrite,
   usePrepareSendTransaction,
   useQuery,
   useSendTransaction,
-  useSigner,
   useWaitForTransaction,
 } from "wagmi";
 import { FaEthereum } from "react-icons/fa";
 import Image from "next/image";
-import { editTrip } from "src/utils/trips";
+
 import { Trip } from "src/utils/interface";
 import { useUser } from "@auth0/nextjs-auth0";
 import { getOrCreateUser } from "src/utils/User";
 import { tokenOptions, validateWeb3Payment } from "src/utils/web3";
-
-import { ArrowDownIcon, ChevronDownIcon } from "@chakra-ui/icons";
-import {
-  Select as ReactSelect,
-  components,
-  SingleValue,
-  SingleValueProps,
-  OptionProps,
-} from "chakra-react-select";
 import { useRouter } from "next/router";
+import { TokenSelect } from "./tokenSelect";
 
 export function SendTransaction({
   value,
@@ -59,29 +38,25 @@ export function SendTransaction({
   setValidate: any;
   validate: string | undefined;
 }) {
+  const addRecentTransaction = useAddRecentTransaction();
+  const [selectedToken, setSelectedToken] = useState(tokenOptions[0]);
+  const priceInEth = (Number(value) / Number(ethPrice)).toFixed(6).toString();
   const { user, isLoading: userLoading } = useUser();
   const { data: userDb, isLoading: userDbLoading } = useQuery(
     ["userDb", user, userLoading],
     () => !userLoading && user && getOrCreateUser(user)
   );
-  const addRecentTransaction = useAddRecentTransaction();
-  const [to, setTo] = useState(address);
-  const [debouncedTo] = useDebounce(to, 500);
-  // const [validate, setValidate] = useState(undefined);
-  const [amount, setAmount] = useState(Number(value).toFixed(6).toString());
-  const [selectedToken, setSelectedToken] = useState(tokenOptions[0]);
-  const [debouncedAmount] = useDebounce(amount, 500);
-  const priceInEth = (Number(value) / Number(ethPrice)).toFixed(6).toString();
-  const router = useRouter();
-  const { config: configWrite } = usePrepareContractWrite({
-    addressOrName: selectedToken.value,
-    contractInterface: erc20ABI,
-    functionName: "transfer",
-    args: [address, parseEther(value)],
-  });
 
+  //Pay with custom token https://wagmi.sh/docs/hooks/useContractWrite
+  const { config: configWrite } = usePrepareContractWrite({
+    addressOrName: selectedToken.value, //contract address
+    contractInterface: erc20ABI, //standard ERC-20 interface https://docs.openzeppelin.com/contracts/4.x/erc20
+    functionName: "transfer", //transfer method from ERC-20 contracts https://docs.ethers.io/v5/single-page/#/v5/api/contract/example/
+    args: [address, parseEther(value)], //[receiver, amount]
+  });
   const { data: dataWrite, write } = useContractWrite(configWrite);
 
+  //Pay with ether https://wagmi.sh/docs/hooks/useSendTransaction
   const { config } = usePrepareSendTransaction({
     request: {
       to: address,
@@ -90,11 +65,13 @@ export function SendTransaction({
   });
   const { data, sendTransaction } = useSendTransaction(config);
 
+  //Wait for payment (ether or custom) to be completed
   const { isLoading, isSuccess } = useWaitForTransaction({
-    hash: data?.hash || dataWrite?.hash,
+    hash: data?.hash || dataWrite?.hash, //transaction hash
   });
+
   const handleTokenChange = (value: any) => {
-    setSelectedToken(value);
+    setSelectedToken(value); //change selected token
   };
 
   const { data: updatedTrip } = useQuery(
@@ -111,53 +88,15 @@ export function SendTransaction({
     () => {
       if (!isLoading && isSuccess && !validate) {
         addRecentTransaction({
+          //add transaction to rainbowkit profile history
           hash: data?.hash || dataWrite?.hash,
           description: tripData.name,
         });
-        validateWeb3Payment(userDb?.data.id, tripData.id);
-        return setValidate("validated");
+        return validateWeb3Payment(userDb?.data.id, tripData.id); //validate payment in database, update trip info
       }
     }
   );
 
-  const DropdownIndicator = (props: any) => {
-    return (
-      <components.DropdownIndicator {...props}>
-        <ChevronDownIcon color={"black"} />
-      </components.DropdownIndicator>
-    );
-  };
-  const Option = (props: any) => (
-    <components.Option {...props}>
-      <Flex color="black">
-        <Image
-          src={props.data.icon}
-          alt="logo"
-          width="30"
-          height="25"
-          // style={{ borderRadius: 10, overflow: "hidden" }}
-        />
-        <Text ml={2}>{props.data.label}</Text>
-      </Flex>
-    </components.Option>
-  );
-
-  const SingleValue = (props: any) => (
-    <components.SingleValue {...props}>
-      <Flex position="absolute" top="15%">
-        <Image
-          src={selectedToken.icon}
-          alt="selected-logo"
-          width="30"
-          height="25"
-          style={{ borderRadius: 10, overflow: "hidden" }}
-        />
-        <Text ml={1} mr={1} fontSize="lg">
-          {props.data.label}
-        </Text>
-      </Flex>
-    </components.SingleValue>
-  );
   return (
     <Box
       w="25.62rem"
@@ -211,7 +150,7 @@ export function SendTransaction({
                   <Input
                     color="black"
                     aria-label="Recipient"
-                    value={to}
+                    value={address}
                     fontWeight="500"
                     fontSize="1.5rem"
                     width="100%"
@@ -233,44 +172,19 @@ export function SendTransaction({
                   border="0.06rem solid rgb(237, 238, 242)"
                   _hover={{ border: "0.06rem solid rgb(211,211,211)" }}
                 >
-                  {/* <Flex
-                  // borderRadius="1.12rem"
-                  // boxShadow="rgba(0, 0, 0, 0.075) 0px 6px 10px"
-                  // fontWeight="500"
-                  // mr="0.5rem"
-                  // color="black"
-                  // // onClick={() => {}}
-                  // _hover={{ bg: "rgb(30,144,255)" }}
-                  > */}
                   <Box
                     w={215}
                     boxShadow="rgb(0 0 0 / 8%) 0rem 0.37rem 0.62rem"
                     borderRadius="1.37rem"
                     bg="#D1DFE3"
                   >
-                    <ReactSelect
-                      value={selectedToken}
-                      options={tokenOptions}
-                      onChange={handleTokenChange}
-                      components={{
-                        Option,
-                        SingleValue,
-                        DropdownIndicator,
-                      }}
+                    <TokenSelect
+                      selectedToken={selectedToken}
+                      tokenOptions={tokenOptions}
+                      handleTokenChange={handleTokenChange}
                     />
                   </Box>
-                  {/* </Flex> */}
-                  {/* <Image
-                      src={"https://i.vgy.me/ThpwD9.png"}
-                      alt={`Ethereum Logo`}
-                      width="80"
-                      height="80"
-                      style={{ borderRadius: 10, overflow: "hidden" }}
-                    />
-                    <Text fontSize={"md"} ml={2} fontWeight={"semibold"}>
-                      ETH
-                    </Text>
-                  </Select> */}
+
                   <Input
                     fontSize="1.5rem"
                     width="100%"
